refactor: refactored algolia logo code
