+++ src/components/icons/index.js
@@ -5,5 +5,5 @@ export { default as GitIcon } from './git';
 export { default as MoonIcon } from './moon';
 export { default as SunIcon } from './sun';
 export { default as SearchIcon } from './search';
-export { default as NoResultIcon } from './no-result-icon';
-export { default as AlgoliaLogo } from './algolia-logo';
+export { default as NoResultIcon } from './no-result';
+export { default as AlgoliaIcon } from './algolia';
