+++ src/components/icons/no-result/index.js
